Reframe Sessions guide as first-run onboarding
