+++ clawjournal/web/frontend/src/components/GettingStartedGuide.tsx
@@ -1,0 +1,128 @@
+import { Link } from 'react-router-dom';
+import type { Stats } from '../types.ts';
+import { colors } from '../theme.ts';
+
+interface GettingStartedGuideProps {
+  stats: Stats;
+  onDismiss: () => void;
+}
+
+export function GettingStartedGuide({ stats, onDismiss }: GettingStartedGuideProps) {
+  const approved = stats.by_status['approved'] ?? 0;
+  const toReview = (stats.by_status['new'] ?? 0) + (stats.by_status['shortlisted'] ?? 0);
+  const steps = [
+    { label: 'Review', detail: toReview > 0 ? `${toReview} waiting` : 'Scan the list' },
+    { label: 'Approve', detail: approved > 0 ? `${approved} approved` : 'Pick traces' },
+    { label: 'Redact', detail: 'Local preview' },
+    { label: 'Share', detail: 'Submit or zip' },
+  ];
+
+  return (
+    <div style={{
+      display: 'flex',
+      flexWrap: 'wrap',
+      gap: 14,
+      alignItems: 'center',
+      margin: '2px 0 12px',
+      padding: '12px 14px',
+      border: `1px solid ${colors.primary200}`,
+      borderRadius: 8,
+      background: colors.primary50,
+    }}>
+      <div style={{ minWidth: 220, flex: '1 1 240px' }}>
+        <div style={{ fontSize: 13.5, fontWeight: 700, color: colors.gray900, marginBottom: 2 }}>
+          New here? Turn your sessions into shareable traces.
+        </div>
+        <div style={{ fontSize: 12, color: colors.gray600, lineHeight: 1.45 }}>
+          Triage your sessions here, then Share redacts and lets you review locally. Nothing leaves your machine until you approve it.
+        </div>
+      </div>
+
+      <div style={{
+        display: 'grid',
+        gridTemplateColumns: 'repeat(auto-fit, minmax(132px, 1fr))',
+        gap: 6,
+        flex: '999 1 440px',
+        minWidth: 280,
+      }}>
+        {steps.map((step, idx) => (
+          <div
+            key={step.label}
+            style={{
+              display: 'flex',
+              gap: 7,
+              alignItems: 'center',
+              minWidth: 0,
+              padding: '7px 8px',
+              border: `1px solid ${colors.gray200}`,
+              borderRadius: 6,
+              background: colors.white,
+            }}
+          >
+            <span style={{
+              width: 20,
+              height: 20,
+              borderRadius: '50%',
+              display: 'inline-grid',
+              placeItems: 'center',
+              flexShrink: 0,
+              background: colors.gray800,
+              color: colors.white,
+              fontSize: 11,
+              fontWeight: 700,
+              fontVariantNumeric: 'tabular-nums',
+            }}>
+              {idx + 1}
+            </span>
+            <span style={{ minWidth: 0 }}>
+              <span style={{ display: 'block', fontSize: 12, fontWeight: 600, color: colors.gray800, whiteSpace: 'nowrap' }}>
+                {step.label}
+              </span>
+              <span style={{ display: 'block', fontSize: 11, color: colors.gray500, lineHeight: 1.25 }}>
+                {step.detail}
+              </span>
+            </span>
+          </div>
+        ))}
+      </div>
+
+      <div style={{ display: 'flex', gap: 8, alignItems: 'center', justifyContent: 'flex-end', marginLeft: 'auto' }}>
+        <Link to="/share" style={{
+          display: 'inline-flex',
+          alignItems: 'center',
+          padding: '7px 12px',
+          background: colors.gray900,
+          color: colors.white,
+          borderRadius: 6,
+          fontSize: 12.5,
+          fontWeight: 600,
+          textDecoration: 'none',
+          whiteSpace: 'nowrap',
+        }}>
+          Open Share
+        </Link>
+        <button
+          type="button"
+          onClick={onDismiss}
+          aria-label="Dismiss"
+          title="Dismiss"
+          style={{
+            width: 28,
+            height: 28,
+            display: 'grid',
+            placeItems: 'center',
+            border: `1px solid ${colors.primary200}`,
+            borderRadius: 6,
+            background: 'transparent',
+            color: colors.gray500,
+            cursor: 'pointer',
+            fontSize: 15,
+            lineHeight: 1,
+          }}
+        >
+          ✕
+        </button>
+      </div>
+    </div>
+  );
+}
