+++ clawjournal/web/frontend/src/views/Inbox.tsx
@@ -6,10 +6,12 @@ import { useToast } from '../components/Toast.tsx';
 import { ConfirmDialog } from '../components/ConfirmDialog.tsx';
 import { Spinner } from '../components/Spinner.tsx';
 import { LABELS } from '../components/BadgeChip.tsx';
+import { GettingStartedGuide } from '../components/GettingStartedGuide.tsx';
 import { colors, selectStyle } from '../theme.ts';
 
 const PAGE_SIZE = 10;
-const SHARE_GUIDE_DISMISSED_KEY = 'cj.shareGuideDismissed';
+const GETTING_STARTED_DISMISSED_KEY = 'cj.gettingStartedDismissed';
+const LEGACY_SHARE_GUIDE_DISMISSED_KEY = 'cj.shareGuideDismissed';
 
 function failureBadge(score: number | null): string {
   if (score == null) return '\u2014';
@@ -102,139 +104,16 @@ function hexAlpha(hex: string, alpha: number): string {
   return `rgba(${r},${g},${b},${alpha})`;
 }
 
-function ShareWorkflowGuide({
-  stats,
-  onDismiss,
-}: {
-  stats: Stats;
-  onDismiss: () => void;
-}) {
-  const approved = stats.by_status['approved'] ?? 0;
-  const toReview = (stats.by_status['new'] ?? 0) + (stats.by_status['shortlisted'] ?? 0);
-  const steps = [
-    { label: 'Review', detail: toReview > 0 ? `${toReview} waiting` : 'Inspect traces' },
-    { label: 'Select', detail: approved > 0 ? `${approved} approved` : 'Choose traces' },
-    { label: 'Redact', detail: 'Before upload' },
-    { label: 'Package', detail: 'Submit or zip' },
-  ];
-
-  return (
-    <div style={{
-      display: 'flex',
-      flexWrap: 'wrap',
-      gap: 14,
-      alignItems: 'center',
-      margin: '2px 0 12px',
-      padding: '12px 14px',
-      border: `1px solid ${colors.primary200}`,
-      borderRadius: 8,
-      background: colors.primary50,
-    }}>
-      <div style={{ minWidth: 220, flex: '1 1 240px' }}>
-        <div style={{ fontSize: 13.5, fontWeight: 700, color: colors.gray900, marginBottom: 2 }}>
-          Start with your local traces
-        </div>
-        <div style={{ fontSize: 12, color: colors.gray600, lineHeight: 1.45 }}>
-          Review traces here. Share redacts and reviews them locally before anything leaves your machine.
-        </div>
-      </div>
-
-      <div style={{
-        display: 'grid',
-        gridTemplateColumns: 'repeat(auto-fit, minmax(132px, 1fr))',
-        gap: 6,
-        flex: '999 1 440px',
-        minWidth: 280,
-      }}>
-        {steps.map((step, idx) => (
-          <div
-            key={step.label}
-            style={{
-              display: 'flex',
-              gap: 7,
-              alignItems: 'center',
-              minWidth: 0,
-              padding: '7px 8px',
-              border: `1px solid ${colors.gray200}`,
-              borderRadius: 6,
-              background: colors.white,
-            }}
-          >
-            <span style={{
-              width: 20,
-              height: 20,
-              borderRadius: '50%',
-              display: 'inline-grid',
-              placeItems: 'center',
-              flexShrink: 0,
-              background: colors.gray800,
-              color: colors.white,
-              fontSize: 11,
-              fontWeight: 700,
-              fontVariantNumeric: 'tabular-nums',
-            }}>
-              {idx + 1}
-            </span>
-            <span style={{ minWidth: 0 }}>
-              <span style={{ display: 'block', fontSize: 12, fontWeight: 600, color: colors.gray800, whiteSpace: 'nowrap' }}>
-                {step.label}
-              </span>
-              <span style={{ display: 'block', fontSize: 11, color: colors.gray500, lineHeight: 1.25 }}>
-                {step.detail}
-              </span>
-            </span>
-          </div>
-        ))}
-      </div>
-
-      <div style={{ display: 'flex', gap: 8, alignItems: 'center', justifyContent: 'flex-end', marginLeft: 'auto' }}>
-        <Link to="/share" style={{
-          display: 'inline-flex',
-          alignItems: 'center',
-          padding: '7px 12px',
-          background: colors.gray900,
-          color: colors.white,
-          borderRadius: 6,
-          fontSize: 12.5,
-          fontWeight: 600,
-          textDecoration: 'none',
-          whiteSpace: 'nowrap',
-        }}>
-          Open Share
-        </Link>
-        <button
-          type="button"
-          onClick={onDismiss}
-          aria-label="Dismiss"
-          title="Dismiss"
-          style={{
-            width: 28,
-            height: 28,
-            display: 'grid',
-            placeItems: 'center',
-            border: `1px solid ${colors.primary200}`,
-            borderRadius: 6,
-            background: 'transparent',
-            color: colors.gray500,
-            cursor: 'pointer',
-            fontSize: 15,
-            lineHeight: 1,
-          }}
-        >
-          &times;
-        </button>
-      </div>
-    </div>
-  );
-}
-
 export function Inbox() {
   const { toast } = useToast();
   const [sessions, setSessions] = useState<Session[]>([]);
   const [stats, setStats] = useState<Stats>({ total: 0, by_status: {}, by_source: {}, by_project: {}, by_task_type: {} });
-  const [showShareGuide, setShowShareGuide] = useState(() => {
+  const [showGettingStartedGuide, setShowGettingStartedGuide] = useState(() => {
     try {
-      return localStorage.getItem(SHARE_GUIDE_DISMISSED_KEY) !== '1';
+      return (
+        localStorage.getItem(GETTING_STARTED_DISMISSED_KEY) !== '1'
+        && localStorage.getItem(LEGACY_SHARE_GUIDE_DISMISSED_KEY) !== '1'
+      );
     } catch {
       return true;
     }
@@ -414,10 +293,14 @@ export function Inbox() {
   const taskTypes = Object.entries(stats.by_task_type ?? {})
     .sort(([, a], [, b]) => b - a);
 
-  const dismissShareGuide = () => {
-    setShowShareGuide(false);
+  const approved = stats.by_status['approved'] ?? 0;
+  const blocked = stats.by_status['blocked'] ?? 0;
+  const hasTriageProgress = approved + blocked > 0;
+
+  const dismissGettingStartedGuide = () => {
+    setShowGettingStartedGuide(false);
     try {
-      localStorage.setItem(SHARE_GUIDE_DISMISSED_KEY, '1');
+      localStorage.setItem(GETTING_STARTED_DISMISSED_KEY, '1');
     } catch { /* ignore */ }
   };
 
@@ -445,8 +328,8 @@ export function Inbox() {
         </div>
       </div>
 
-      {showShareGuide && stats.total > 0 && sessions.length > 0 && !typeFilter && (
-        <ShareWorkflowGuide stats={stats} onDismiss={dismissShareGuide} />
+      {showGettingStartedGuide && stats.total > 0 && sessions.length > 0 && !hasTriageProgress && !typeFilter && (
+        <GettingStartedGuide stats={stats} onDismiss={dismissGettingStartedGuide} />
       )}
 
       {showFilters && (
